+++ bcos-rpc/test/CMakeLists.txt
@@ -40,7 +40,8 @@ set_source_files_properties("unittests/rpc/RpcReadHandlersTest.cpp" PROPERTIES S
 # FullChainFixture.h pulls RocksDB and testutils headers whose names collide with the
 # using-directives other unity-merged TUs carry; compile its user standalone.
 set_source_files_properties(
-    "unittests/rpc/EthGetProofIntegrationTest.cpp" PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
+    "unittests/rpc/EthGetProofIntegrationTest.cpp" "unittests/rpc/FullChainFixtureInst.cpp"
+    PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
 set_target_properties(${TEST_BINARY_NAME} PROPERTIES UNITY_BUILD "ON")
 #add_test(NAME ${TEST_BINARY_NAME} WORKING_DIRECTORY ${CMAKE_RUNTIME_OUTPUT_DIRECTORY} COMMAND ${TEST_BINARY_NAME})
 config_test_cases("" "${SOURCES}" ${TEST_BINARY_NAME} "" "")
