+++ transaction-scheduler/tests/CMakeLists.txt
@@ -28,7 +28,8 @@ set_source_files_properties("main.cpp" PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
 # using-directives other unity-merged TUs carry; compile its users standalone.
 set_source_files_properties(
     "FullChainFixtureSmokeTest.cpp" "BaselineSchedulerMPTL1UpgradeTest.cpp"
-    "BaselineSchedulerMPTGenesisTest.cpp" PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
+    "BaselineSchedulerMPTGenesisTest.cpp" "SharedBaselineSchedulerInst.cpp"
+    PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
 # TestEthereumExecutorScheduler pulls EthereumExecutor + bcos-ledger LedgerMethods.h
 # (executor / tars headers) whose namespace-scope entities would collide with the
 # other unity-merged TUs; compile standalone.
