+++ transaction-scheduler/tests/FullChainFixture.h
@@ -480,3 +480,11 @@ public:
 };
 
 }  // namespace bcos::test::fullchain
+
+// Explicitly instantiated once per test binary (SharedBaselineSchedulerInst.cpp in
+// test-transaction-scheduler, unittests/rpc/FullChainFixtureInst.cpp in test-bcos-rpc)
+// so each consumer TU — the MPT genesis / L1-upgrade / smoke tests — parses only the
+// declaration header instead of instantiating the whole scheduler.
+extern template class bcos::scheduler_v1::BaselineScheduler<bcos::test::fullchain::FCMultiLayerStorage,
+    bcos::test::fullchain::FCExecutor, bcos::test::fullchain::FCWritingScheduler,
+    bcos::ledger::Ledger>;
