+++ transaction-scheduler/tests/TestExecuteStateRootRegression.cpp
@@ -38,7 +38,7 @@
 #include "bcos-tars-protocol/protocol/TransactionReceiptFactoryImpl.h"
 #include "bcos-tars-protocol/protocol/TransactionReceiptImpl.h"
 #include "bcos-task/AwaitableValue.h"
-#include "bcos-transaction-scheduler/BaselineScheduler.h"
+#include "SharedBaselineSchedulerMock.h"
 #include <boost/test/unit_test.hpp>
 #include <array>
 #include <fakeit.hpp>
@@ -52,11 +52,11 @@
 // argument, so no existing test could see the reroute — the scheduler impl
 // here does write, which is the whole point of this file.
 //
-// Everything lives in an anonymous namespace (unity-build friendly), and the
-// storage stack uses its own backend hasher type so this TU's ViewType — and
-// therefore its getLedgerConfig tag_invoke stub — cannot collide with the
-// stubs in testBaselineScheduler.cpp / FIB101_102_103_104_SchedulerTest.cpp
-// when CMake merges the sources into one unity TU.
+// Everything lives in an anonymous namespace (unity-build friendly). The storage stack,
+// mock executor / scheduler and the getLedgerConfig stub are the SHARED test mocks
+// (SharedBaselineSchedulerMock.h): the old WritingMockScheduler's per-block row writes
+// became SharedMockScheduler::m_everyBlockRows, and the whole-suite BaselineScheduler
+// instantiation happens once in SharedBaselineSchedulerInst.cpp.
 namespace
 {
 using namespace bcos;
@@ -64,16 +64,9 @@ using namespace bcos::storage2;
 using namespace bcos::executor_v1;
 using namespace bcos::scheduler_v1;
 
-struct SRBackendHash : std::hash<StateKey>
-{
-};
-
-using SRMutableStorage = memory_storage::MemoryStorage<StateKey, StateValue,
-    memory_storage::Attribute(memory_storage::ORDERED | memory_storage::LOGICAL_DELETION)>;
-using SRBackendStorage = memory_storage::MemoryStorage<StateKey, StateValue,
-    memory_storage::Attribute(memory_storage::ORDERED | memory_storage::CONCURRENT), SRBackendHash>;
-using SRCheckpointBackend = TrivialCheckpointStorage<StateKey, StateValue, SRBackendStorage>;
-using SRMultiLayerStorage = MultiLayerStorage<SRMutableStorage, void, SRCheckpointBackend>;
+using bcos::test::sharedmock::SharedBackendStorage;
+using bcos::test::sharedmock::SharedCheckpointBackend;
+using bcos::test::sharedmock::SharedMultiLayerStorage;
 
 struct StateRow
 {
@@ -85,77 +78,6 @@ constexpr std::array<StateRow, 2> stateRows{
     StateRow{"/apps/aabbccdd00112233", "balance", "1000000"},
     StateRow{"/apps/eeff445566778899", "code", "0x60806040526004361061"},
 };
-
-struct SRMockExecutor
-{
-    task::Task<protocol::TransactionReceipt::Ptr> executeTransaction(auto& storage,
-        protocol::BlockHeader const& blockHeader, protocol::Transaction const& transaction,
-        int contextID, ledger::LedgerConfig const& ledgerConfig, bool call)
-    {
-        co_return {};
-    }
-
-    template <class Storage>
-    struct ExecuteContext
-    {
-        task::Task<void> prepare() { co_return; }
-        task::Task<void> execute() { co_return; }
-        task::Task<protocol::TransactionReceipt::Ptr> finish() { co_return {}; }
-    };
-
-    auto createExecuteContext(auto& storage, protocol::BlockHeader const& blockHeader,
-        protocol::Transaction const& transaction, int32_t contextID,
-        ledger::LedgerConfig const& ledgerConfig, bool call)
-        -> task::Task<ExecuteContext<std::decay_t<decltype(storage)>>>
-    {
-        co_return {};
-    }
-};
-
-// Unlike the stock MockScheduler, this impl writes state rows through the
-// storage argument BaselineScheduler hands it — exactly what the production
-// SchedulerSerialImpl / SchedulerParallelImpl do. If coExecuteBlock passes the
-// backend instead of the layered view, these rows skip the mutable layer and
-// the stateRoot XOR collapses to zero.
-struct WritingMockScheduler
-{
-    task::Task<std::vector<protocol::TransactionReceipt::Ptr>> executeBlock(auto& storage,
-        auto& executor, protocol::BlockHeader const& blockHeader,
-        ::ranges::input_range auto const& transactions, ledger::LedgerConfig const& /*unused*/)
-    {
-        for (const auto& row : stateRows)
-        {
-            storage::Entry entry;
-            entry.set(row.value);
-            co_await storage2::writeOne(storage, StateKey{row.table, row.key}, std::move(entry));
-        }
-
-        auto receipts =
-            ::ranges::iota_view<size_t, size_t>(0, ::ranges::size(transactions)) |
-            ::ranges::views::transform([](size_t index) -> protocol::TransactionReceipt::Ptr {
-                auto receipt = std::make_shared<bcostars::protocol::TransactionReceiptImpl>();
-                constexpr static std::string_view str = "abc";
-                auto& inner = receipt->inner();
-                inner.dataHash.assign(str.begin(), str.end());
-                inner.data.gasUsed = "100";
-                return receipt;
-            }) |
-            ::ranges::to<std::vector<protocol::TransactionReceipt::Ptr>>();
-
-        co_return receipts;
-    }
-};
-
-// Storage-level getLedgerConfig stub, found via ADL when the BaselineScheduler
-// template is instantiated; [[maybe_unused]] silences clang's -Wunused-function
-// false positive on indirect template use.
-[[maybe_unused]] task::AwaitableValue<void> tag_invoke(
-    ledger::tag_t<bcos::ledger::getLedgerConfig> /*unused*/,
-    SRMultiLayerStorage::ViewType& /*storage*/, bcos::ledger::LedgerConfig& /*ledgerConfig*/,
-    protocol::BlockNumber /*blockNumber*/, protocol::BlockFactory& /*blockFactory*/)
-{
-    return {};
-}
 
 task::Task<std::vector<protocol::Transaction::ConstPtr>> emptyTxsTaskSR()
 {
@@ -193,6 +115,17 @@ public:
         baselineScheduler(multiLayerStorage, mockScheduler, mockExecutor, *blockFactory,
             mockLedger.get(), mockTxPool.get(), *transactionSubmitResultFactory, *hashImpl)
     {
+        // Shared mock configuration (SharedBaselineSchedulerMock.h): the two state rows
+        // the old WritingMockScheduler wrote on EVERY block; trivial getLedgerConfig
+        // stub behaviour (default Features).
+        for (auto const& row : stateRows)
+        {
+            mockScheduler.m_everyBlockRows.push_back(
+                bcos::test::sharedmock::SharedMockScheduler::RowOp{
+                    std::string{row.table}, std::string{row.key}, std::string{row.value}});
+        }
+        bcos::test::sharedmock::g_stubFeatures = ledger::Features{};
+
         fakeit::When(Method(mockLedger, asyncPrewriteBlock))
             .AlwaysDo([](storage::StorageInterface::Ptr, protocol::ConstTransactionsPtr,
                           protocol::Block::ConstPtr,
@@ -299,8 +232,8 @@ public:
 
     static constexpr uint32_t blockVersion = 200;
 
-    SRBackendStorage backendStorage;
-    SRCheckpointBackend checkpointBackend;
+    SharedBackendStorage backendStorage;
+    SharedCheckpointBackend checkpointBackend;
     crypto::CryptoSuite::Ptr cryptoSuite;
     std::shared_ptr<bcostars::protocol::BlockHeaderFactoryImpl> blockHeaderFactory;
     std::shared_ptr<bcostars::protocol::TransactionFactoryImpl> transactionFactory;
@@ -310,14 +243,12 @@ public:
 
     crypto::Hash::Ptr hashImpl = std::make_shared<crypto::Keccak256>();
 
-    WritingMockScheduler mockScheduler;
+    bcos::test::sharedmock::SharedMockScheduler mockScheduler;
     fakeit::Mock<ledger::LedgerInterface> mockLedger;
     fakeit::Mock<txpool::TxPoolInterface> mockTxPool;
-    SRMultiLayerStorage multiLayerStorage;
-    SRMockExecutor mockExecutor;
-    BaselineScheduler<decltype(multiLayerStorage), SRMockExecutor, WritingMockScheduler,
-        ledger::LedgerInterface>
-        baselineScheduler;
+    SharedMultiLayerStorage multiLayerStorage;
+    bcos::test::sharedmock::SharedMockExecutor mockExecutor;
+    bcos::test::sharedmock::SharedBaselineScheduler baselineScheduler;
 };
 
 BOOST_FIXTURE_TEST_SUITE(TestExecuteStateRootRegression, StateRootRegressionFixture)
