+++ transaction-scheduler/tests/testBaselineScheduler.cpp
@@ -18,7 +18,7 @@
 #include "bcos-tars-protocol/protocol/TransactionReceiptFactoryImpl.h"
 #include "bcos-tars-protocol/protocol/TransactionReceiptImpl.h"
 #include "bcos-task/AwaitableValue.h"
-#include "bcos-transaction-scheduler/BaselineScheduler.h"
+#include "SharedBaselineSchedulerMock.h"
 #include <boost/test/unit_test.hpp>
 #include <fakeit.hpp>
 #include <future>
@@ -28,81 +28,9 @@ using namespace bcos::storage2;
 using namespace bcos::executor_v1;
 using namespace bcos::scheduler_v1;
 
-using MutableStorage = memory_storage::MemoryStorage<StateKey, StateValue,
-    memory_storage::Attribute(memory_storage::ORDERED | memory_storage::LOGICAL_DELETION)>;
-using BackendStorage = memory_storage::MemoryStorage<StateKey, StateValue,
-    memory_storage::Attribute(memory_storage::ORDERED | memory_storage::CONCURRENT),
-    std::hash<StateKey>>;
-using CheckpointBackend = TrivialCheckpointStorage<StateKey, StateValue, BackendStorage>;
-using MyMultiLayerStorage = MultiLayerStorage<MutableStorage, void, CheckpointBackend>;
-
-struct MockExecutorBaseline
-{
-    task::Task<protocol::TransactionReceipt::Ptr> executeTransaction(auto& storage,
-        protocol::BlockHeader const& blockHeader, protocol::Transaction const& transaction,
-        int contextID, ledger::LedgerConfig const& ledgerConfig, bool call)
-    {
-        if (transaction.version() == 99)
-        {
-            BOOST_CHECK_EQUAL(blockHeader.timestamp(), 10088);
-        }
-        co_return {};
-    }
-
-    template <class Storage>
-    struct ExecuteContext
-    {
-        task::Task<void> prepare() { co_return; }
-        task::Task<void> execute() { co_return; }
-        task::Task<protocol::TransactionReceipt::Ptr> finish() { co_return {}; }
-    };
-
-    auto createExecuteContext(auto& storage, protocol::BlockHeader const& blockHeader,
-        protocol::Transaction const& transaction, int32_t contextID,
-        ledger::LedgerConfig const& ledgerConfig, bool call)
-        -> task::Task<ExecuteContext<std::decay_t<decltype(storage)>>>
-    {
-        co_return {};
-    }
-};
-struct MockScheduler
-{
-    task::Task<std::vector<protocol::TransactionReceipt::Ptr>> executeBlock(auto& storage,
-        auto& executor, protocol::BlockHeader const& blockHeader,
-        ::ranges::input_range auto const& transactions, ledger::LedgerConfig const& /*unused*/)
-    {
-        auto receipts =
-            ::ranges::iota_view<size_t, size_t>(0, ::ranges::size(transactions)) |
-            ::ranges::views::transform([](size_t index) -> protocol::TransactionReceipt::Ptr {
-                auto receipt = std::make_shared<bcostars::protocol::TransactionReceiptImpl>();
-                constexpr static std::string_view str = "abc";
-                auto& inner = receipt->inner();
-                inner.dataHash.assign(str.begin(), str.end());
-                inner.data.gasUsed = "100";
-
-                bytes logAddress;
-                logAddress.assign(str.begin(), str.end());
-                bcos::protocol::LogEntry logEntry{
-                    logAddress, bcos::h256s{bcos::h256{}}, bcos::bytes{}};
-                std::vector<bcos::protocol::LogEntry> logs;
-                logs.emplace_back(std::move(logEntry));
-                receipt->setLogEntries(logs);
-                return receipt;
-            }) |
-            ::ranges::to<std::vector<protocol::TransactionReceipt::Ptr>>();
-
-        co_return receipts;
-    }
-};
-
-// Keep storage-level getLedgerConfig stub minimal for tests
-inline task::AwaitableValue<void> tag_invoke(
-    ledger::tag_t<bcos::ledger::getLedgerConfig> /*unused*/, MyMultiLayerStorage::ViewType& storage,
-    bcos::ledger::LedgerConfig& ledgerConfig, protocol::BlockNumber blockNumber,
-    protocol::BlockFactory& blockFactory)
-{
-    return {};
-}
+using bcos::test::sharedmock::SharedBackendStorage;
+using bcos::test::sharedmock::SharedCheckpointBackend;
+using bcos::test::sharedmock::SharedMultiLayerStorage;
 
 // Helper: empty task returning no transactions
 static bcos::task::Task<std::vector<bcos::protocol::Transaction::ConstPtr>> emptyTxsTask()
@@ -131,6 +59,13 @@ public:
         baselineScheduler(multiLayerStorage, mockScheduler, mockExecutor, *blockFactory,
             mockLedger.get(), mockTxPool.get(), *transactionSubmitResultFactory, *hashImpl)
     {
+        // Shared mock configuration (SharedBaselineSchedulerMock.h): the version-99
+        // timestamp probe and log-carrying receipts this file's mocks used to provide;
+        // the trivial getLedgerConfig stub behaviour (default Features).
+        mockExecutor.m_checkTimestamp99 = true;
+        mockScheduler.m_receiptsWithLogs = true;
+        bcos::test::sharedmock::g_stubFeatures = ledger::Features{};
+
         // Ledger: asyncPrewriteBlock => invoke callback(success)
         fakeit::When(Method(mockLedger, asyncPrewriteBlock))
             .AlwaysDo([](bcos::storage::StorageInterface::Ptr, bcos::protocol::ConstTransactionsPtr,
@@ -184,20 +119,18 @@ public:
     std::shared_ptr<bcostars::protocol::BlockFactoryImpl> blockFactory;
     std::shared_ptr<protocol::TransactionSubmitResultFactoryImpl> transactionSubmitResultFactory;
 
-    BackendStorage backendStorage;
-    CheckpointBackend checkpointBackend;
+    SharedBackendStorage backendStorage;
+    SharedCheckpointBackend checkpointBackend;
 
     crypto::Hash::Ptr hashImpl = std::make_shared<bcos::crypto::Keccak256>();
 
-    MockScheduler mockScheduler;
+    bcos::test::sharedmock::SharedMockScheduler mockScheduler;
     // FakeIt mocks for ledger and txpool
     fakeit::Mock<bcos::ledger::LedgerInterface> mockLedger;
     fakeit::Mock<bcos::txpool::TxPoolInterface> mockTxPool;
-    MyMultiLayerStorage multiLayerStorage;
-    MockExecutorBaseline mockExecutor;
-    BaselineScheduler<decltype(multiLayerStorage), MockExecutorBaseline, MockScheduler,
-        bcos::ledger::LedgerInterface>
-        baselineScheduler;
+    SharedMultiLayerStorage multiLayerStorage;
+    bcos::test::sharedmock::SharedMockExecutor mockExecutor;
+    bcos::test::sharedmock::SharedBaselineScheduler baselineScheduler;
 };
 
 BOOST_FIXTURE_TEST_SUITE(TestBaselineScheduler, TestBaselineSchedulerFixture)
